+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.21-1.1.1]
+### Changed
+- Barley block renamed to "Barley Straw Block" (ID is still `berriesandmore:barley_block`)
+### Fixed
+- All barley blocks and items (except orzo) can now be composted
+
 ## [1.21-1.1.0]
 - Updated to Minecraft 1.21 (Neoforge)
 ### Added